[Fuzzing] Move transaction related fuzzers into single file and rename fuzz target files to be more obvious.
