+++ testsuite/libra-fuzzer/src/fuzz_targets.rs
@@ -7,37 +7,36 @@ use once_cell::sync::Lazy;
 use std::{collections::BTreeMap, env};
 
 // List fuzz target modules here.
-mod compiled_module;
-mod consensus_proposal;
-mod execute_and_commit_chunk;
-mod inbound_rpc_protocol;
-mod inner_signed_transaction;
+mod consensus;
+mod executor;
 mod json_rpc_service;
-mod language_transaction_execution;
 mod mempool;
+mod move_vm;
+mod network;
 mod network_noise;
-mod state_sync_msg;
+mod state_sync;
 mod storage;
-mod vm_value;
+mod transaction;
+mod vm;
 
 static ALL_TARGETS: Lazy<BTreeMap<&'static str, Box<dyn FuzzTargetImpl>>> = Lazy::new(|| {
     let targets: Vec<Box<dyn FuzzTargetImpl>> = vec![
         // List fuzz targets here in this format.
-        Box::new(compiled_module::CompiledModuleTarget::default()),
-        Box::new(consensus_proposal::ConsensusProposal::default()),
-        Box::new(inbound_rpc_protocol::RpcInboundRequest::default()),
-        Box::new(inner_signed_transaction::SignedTransactionTarget::default()),
+        Box::new(vm::CompiledModuleTarget::default()),
+        Box::new(consensus::ConsensusProposal::default()),
+        Box::new(network::RpcInboundRequest::default()),
+        Box::new(transaction::SignedTransactionTarget::default()),
         Box::new(json_rpc_service::JsonRpcSubmitTransactionRequest::default()),
         Box::new(mempool::MempoolIncomingTransactions::default()),
         Box::new(network_noise::NetworkNoiseInitiator::default()),
         Box::new(network_noise::NetworkNoiseResponder::default()),
         Box::new(network_noise::NetworkNoiseStream::default()),
-        Box::new(state_sync_msg::StateSyncMsg::default()),
-        Box::new(language_transaction_execution::LanguageTransactionExecution::default()),
+        Box::new(state_sync::StateSyncMsg::default()),
+        Box::new(transaction::LanguageTransactionExecution::default()),
         //        Box::new(storage_save_blocks::StorageSaveBlocks::default()),
         Box::new(storage::StorageSchemaDecode::default()),
-        Box::new(vm_value::ValueTarget::default()),
-        Box::new(execute_and_commit_chunk::ExecuteAndCommitChunk::default()),
+        Box::new(move_vm::ValueTarget::default()),
+        Box::new(executor::ExecuteAndCommitChunk::default()),
     ];
     targets
         .into_iter()
