+++ testsuite/libra-fuzzer/src/fuzz_targets/transaction.rs
@@ -6,7 +6,8 @@ use language_e2e_tests::account_universe::{
     all_transactions_strategy, log_balance_strategy, run_and_assert_universe, AccountUniverseGen,
 };
 use libra_proptest_helpers::ValueGenerator;
-use proptest::{collection::vec, test_runner};
+use libra_types::transaction::SignedTransaction;
+use proptest::{collection::vec, prelude::*, test_runner};
 
 #[derive(Clone, Debug, Default)]
 pub struct LanguageTransactionExecution;
@@ -31,5 +32,23 @@ impl FuzzTargetImpl for LanguageTransactionExecution {
         let universe = generator.generate(universe_strategy);
 
         run_and_assert_universe(universe, txns).unwrap();
+    }
+}
+
+#[derive(Clone, Debug, Default)]
+pub struct SignedTransactionTarget;
+
+impl FuzzTargetImpl for SignedTransactionTarget {
+    fn description(&self) -> &'static str {
+        "SignedTransaction (LCS deserializer)"
+    }
+
+    fn generate(&self, _idx: usize, gen: &mut ValueGenerator) -> Option<Vec<u8>> {
+        let value = gen.generate(any_with::<SignedTransaction>(()));
+        Some(lcs::to_bytes(&value).expect("serialization should work"))
+    }
+
+    fn fuzz(&self, data: &[u8]) {
+        let _: Result<SignedTransaction, _> = lcs::from_bytes(&data);
     }
 }
