+++ testsuite/libra-fuzzer/src/fuzz_targets/vm.rs
